+++ packages/cli/src/command-utils/common-types.ts
@@ -1,15 +1,4 @@
-import {
-  ReplayExecutionOptions,
-  ReplayTarget,
-  ScreenshottingEnabledOptions,
-} from "@alwaysmeticulous/common";
-import defaults from "lodash/defaults";
-import { TestCaseReplayOptions } from "../config/config.types";
-
-export interface ScreenshotDiffOptions {
-  diffThreshold: number;
-  diffPixelThreshold: number;
-}
+import { ScreenshottingEnabledOptions } from "@alwaysmeticulous/common";
 
 /**
  * Options for taking a screenshot and comparing it against a previous screenshot
@@ -23,57 +12,7 @@ export interface ScreenshotAssertionsEnabledOptions
   diffOptions: ScreenshotDiffOptions;
 }
 
-export const getReplayTarget = ({
-  appUrl,
-  simulationIdForAssets,
-}: {
-  appUrl?: string | undefined;
-  simulationIdForAssets?: string | undefined;
-}): ReplayTarget => {
-  if (simulationIdForAssets) {
-    return { type: "snapshotted-assets", simulationIdForAssets };
-  }
-  if (appUrl) {
-    return { type: "url", appUrl };
-  }
-  return { type: "original-recorded-url" };
-};
-
-export const applyTestCaseExecutionOptionOverrides = (
-  executionOptionsFromCliFlags: ReplayExecutionOptions,
-  overridesFromTestCase: TestCaseReplayOptions
-) => {
-  // Options specified in the test case override those passed as CLI flags
-  // (CLI flags set the defaults)
-  return defaults(
-    {},
-    {
-      moveBeforeClick: overridesFromTestCase.moveBeforeClick,
-      simulationIdForAssets: overridesFromTestCase.simulationIdForAssets,
-    },
-    executionOptionsFromCliFlags
-  );
-};
-
-export const applyTestCaseScreenshottingOptionsOverrides = (
-  screenshottingOptionsFromCliFlags: ScreenshotAssertionsEnabledOptions,
-  overridesFromTestCase: TestCaseReplayOptions
-): ScreenshotAssertionsEnabledOptions => {
-  // Options specified in the test case override those passed as CLI flags
-  // (CLI flags set the defaults)
-  const diffOptions: ScreenshotDiffOptions = defaults(
-    {},
-    {
-      diffThreshold: overridesFromTestCase.diffThreshold,
-      diffPixelThreshold: overridesFromTestCase.diffPixelThreshold,
-    },
-    screenshottingOptionsFromCliFlags.diffOptions
-  );
-  return {
-    enabled: true,
-    screenshotSelector:
-      overridesFromTestCase.screenshotSelector ??
-      screenshottingOptionsFromCliFlags.screenshotSelector,
-    diffOptions,
-  };
-};
+export interface ScreenshotDiffOptions {
+  diffThreshold: number;
+  diffPixelThreshold: number;
+}
